add and delete some files
