+++ python/src/CMakeLists.txt
@@ -227,8 +227,8 @@ ot_add_python_module(statistics statistics_module.i
                    
 ot_add_python_module(iterative_statistics iterative_statistics_module.i
                       IterativeAlgorithm.i IterativeAlgorithm_doc.i.in
+                      IterativAlgorithmImplementation.i IterativeAlgorithmImplementation_doc.i.in
                       MelissaStudy.i MelissaStudy_doc.i.in
-                      IterativeMeanImplementation.i IterativeMeanImplementation_doc.i.in
                       IterativeMean.i IterativeMean_doc.i.in
                       BaseIterativeStatisticsTemplateDefs.i
                    )
