--- python/src/IterativeMeanImplementation_doc.i.in
@@ -1,1 +1,0 @@
-
